move constants to src folder
