+++ src/components/idleUsers/card/index.tsx
@@ -5,7 +5,7 @@ import { UserStatus } from '@/interfaces/userStatus.type';
 import getIdleSinceText from '@/helperFunctions/getIdleSinceText';
 import styles from '@/components/idleUsers/card/card.module.scss';
 import { DUMMY_PROFILE } from '@/components/constants/display-sections.js';
-import { MEMBERS_URL } from '@/components/constants/url';
+import { MEMBERS_URL } from '@/constants/url';
 
 type Props = {
     user: UserStatus;
