+++ src/components/navBar/index.tsx
@@ -13,7 +13,7 @@ import {
     STATUS_URL,
     GITHUB_LOGO,
     RDS_LOGO,
-} from '@/components/constants/url';
+} from '@/constants/url';
 import Dropdown from '../Dropdown/Dropdown';
 import styles from '@/components/navBar/navBar.module.scss';
 import { useGetUserQuery } from '@/app/services/userApi';
