+++ src/constants/display-sections.js
