+++ src/context/isUserAuthorized.tsx
@@ -8,7 +8,7 @@ import {
 } from 'react';
 import fetch from '@/helperFunctions/fetch';
 import { toast, ToastTypes } from '@/helperFunctions/toast';
-import { USER_SELF } from '@/components/constants/url';
+import { USER_SELF } from '@/constants/url';
 
 const { ERROR } = ToastTypes;
 export const isUserAuthorizedContext = createContext<boolean>(false);
