+++ src/pages/mine.tsx
@@ -6,7 +6,7 @@ import Card from '@/components/tasks/card';
 import useFetch from '@/hooks/useFetch';
 import classNames from '@/styles/tasks.module.scss';
 import task from '@/interfaces/task.type';
-import { LOGIN_URL, MINE_TASKS_URL } from '@/components/constants/url';
+import { LOGIN_URL, MINE_TASKS_URL } from '@/constants/url';
 
 function CardList(tasks: task[]) {
     return tasks.map((item: task) => (
